refactor(encoding): extract DfastMatchGenerator into dfast/ module
Move the `DfastMatchGenerator` struct, its inherent impl, and the
dfast-specific constants out of `match_generator.rs` into a dedicated
`dfast::` submodule (~620 LoC). Mechanical move — names, fields,
method bodies, and the `#[inline]` annotations are preserved; the
visibility on the relocated items is opened to `pub(crate)` so:

- `MatchGeneratorDriver` in `match_generator` can keep instantiating
  the dfast backend, configuring `hash_bits` / `lazy_depth` /
  `use_fast_loop`, and accessing the hash tables during reset /
  configure paths.
- The encoder integration tests can keep inspecting internal state
  (`dfast.hash4`, `dfast.hash8`, `dfast.seed_remaining_hashable_starts`,
  the dispatch-only fields needed by driver-switch tests).

`match_generator.rs` shrinks by another 624 lines (9462 → 8838 LoC).
The DFAST_* constants (`DFAST_MIN_MATCH_LEN`, `DFAST_HASH_BITS`, …,
`DFAST_INCOMPRESSIBLE_SKIP_STEP`) and `MIN_WINDOW_LOG` are promoted to
`pub(crate)` so the relocated matcher can keep referencing them. The
`block_looks_incompressible{,_strict}` and `match_table::helpers::*`
imports were only needed by Dfast/Row callers in the monolith and are
now dropped from `match_generator`'s use list (still imported in the
two extracted modules that consume them).

Build, lint (x86_64 + aarch64), and the level22 ratio gate stay green
on 89/89 unit + integration tests.

Part of #111 Phase 1b.
